refactor(client-3d): extract JukeboxRoom props to scene/props/
Phase 3A (partial): split JukeboxRoom.tsx 1455→417 LOC by moving
diner-themed inline prop components into scene/props/diner/ and
single-concept props into scene/props/. Room file now focuses on
geometry, wall occlusion, and composition; props are independently
loadable for agent edits.
